refactor(mitm-addon): extract terminal usage lifecycle
